keep firmware
The system may depend on special firmware in /lib/firmware, so keep it.
Automatically emerging sys-kernel/linux-firmware would also install
firmware. But who knows that contain everything we need. For
instance a 4.9 kernel relies on very recent firmware for intel wifi
which was not in the stable linux-firmware package yet. So leave that
up to the user.